refactor: extract DateTimeFormatParser helper
