+++ src/reuse/modules/util/formatter.ts
@@ -1,8 +1,9 @@
 "use strict";
 
 import { DateFormats, TimeFormats } from "./constants/formatter.constants";
-import { DateFormatsType, DateTimeFormatsType } from "./types/formatter.types";
+import { DateFormatsType, DateTimeFormatsType, TimeFormatsType } from "./types/formatter.types";
 import ErrorHandler from "../../helper/errorHandler";
+import { DateTimeFormatParser } from "../../helper/dateTimeFormatParser";
 
 /**
  * @class formatter
@@ -237,29 +238,21 @@ export class Formatter {
   }
 
   formatDateWithTime(date: Date, format: DateTimeFormatsType = DateFormats.OBJECT, locale = "en-US"): string | Date {
-    const dateFormat = this._extractDateFormat(format);
+    const dateFormat = DateTimeFormatParser.extractDateFormat(format);
     this._validateDateFormat(format, dateFormat);
     const dateFormatted = this.formatDate(date, dateFormat, locale);
     if (dateFormat === DateFormats.DATETIME) {
       return dateFormatted;
     }
 
-    const timeFormat = this._extractTimeFormat(format, dateFormat);
+    const timeFormat = DateTimeFormatParser.extractTimeFormat(format);
     const timeFormatted = this._formatTime(date, timeFormat);
 
-    const delimiter = this._extractDelimiter(format, dateFormat, timeFormat);
+    const delimiter = DateTimeFormatParser.extractDelimiter({ format, dateFormat, timeFormat });
     return `${dateFormatted}${delimiter}${timeFormatted}`;
   }
 
   // =================================== HELPER ===================================
-  private _extractDateFormat(format: DateTimeFormatsType): DateFormatsType {
-    const dateFormat = Object.values(DateFormats).find((f) => format.startsWith(f));
-    if (!dateFormat) {
-      throw new Error("Invalid date format provided.");
-    }
-    return dateFormat as DateFormatsType;
-  }
-
   private _validateDateFormat(format: DateTimeFormatsType, dateFormat?: DateFormatsType): void {
     if (dateFormat === DateFormats.OBJECT && format.toString() !== DateFormats.OBJECT
       || dateFormat === DateFormats.DATETIME && format.toString() !== DateFormats.DATETIME
@@ -268,19 +261,7 @@ export class Formatter {
     }
   }
 
-  private _extractTimeFormat(format: DateTimeFormatsType, dateFormat: DateFormatsType): TimeFormats {
-    const timeFormat = Object.values(TimeFormats).find((f) => format.endsWith(f));
-    if (!timeFormat) {
-      throw new Error("Invalid time format provided.");
-    }
-    return timeFormat as TimeFormats;
-  }
-
-  private _extractDelimiter(format: DateTimeFormatsType, dateFormat: DateFormatsType, timeFormat: TimeFormats): string {
-    return format.slice(dateFormat.length, -timeFormat.length);
-  }
-
-  private _formatTime(date: Date, format: TimeFormats): string {
+  private _formatTime(date: Date, format: TimeFormatsType): string {
     return format.replace(/HH|h|mm|ss|a|z/g, (token) => {
       switch (token) {
         case "HH":
